Unit tests for dkan_api (#105)

+++ modules/custom/dkan_api/src/Storage/DrupalNodeDataset.php
@@ -2,60 +2,107 @@
 
 namespace Drupal\dkan_api\Storage;
 
-use Drupal\node\Entity\Node;
-use Contracts\Storage;
-use Contracts\BulkRetriever;
+use Harvest\Storage\Storage;
+use Drupal\Core\Entity\EntityTypeManagerInterface;
+use Drupal\dkan_api\Storage\ThemeValueReferencer;
 
-class DrupalNodeDataset implements Storage, BulkRetriever {
+/**
+ * DrupalNodeDataset.
+ */
+class DrupalNodeDataset implements Storage {
 
+  /**
+   * Entity Type Manager.
+   *
+   * @var \Drupal\Core\Entity\EntityTypeManagerInterface
+   */
+  protected $entityTypeManager;
+  
   /**
    * @var Drupal\dkan_api\Storage\ThemeValueReferencer
    */
-  private $themeValueReferencer;
+  protected $themeValueReferencer;
 
   /**
-   * Constructs a DrupalNodeDataset.
+   * Constructor.
+   *
+   * @param EntityTypeManagerInterface $entityTypeManager
+   *   Injected entity type manager.
+   * @param ThemeValueReferencer $themeValueReferencer
+   *   Injected theme value referencer.
    */
-  public function __construct() {
-    $this->themeValueReferencer = new ThemeValueReferencer();
+  public function __construct(
+          EntityTypeManagerInterface $entityTypeManager,
+          ThemeValueReferencer $themeValueReferencer
+    ) {
+    $this->entityTypeManager = $entityTypeManager;
+    $this->themeValueReferencer = $themeValueReferencer;
   }
 
+  /**
+   * Get the node storage.
+   *
+   * @return \Drupal\node\NodeStorageInterface
+   *   Node Storage.
+   */
+  protected function getNodeStorage() {
+    return $this->entityTypeManager
+      ->getStorage('node');
+  }
+
+  /**
+   * @return string
+   *   Type of node.
+   */
   protected function getType() {
     return 'data';
   }
 
+  /**
+   * {@inheritDoc}.
+   */
   public function retrieve(string $id): ?string {
 
-    foreach ($this->getNodesByUuid($id) as $nid) {
-      $node = Node::load($nid);
+    if (FALSE !== ($node = $this->getNodeByUuid($id))) {
       return $this->themeDereferenced($node->field_json_metadata->value);
     }
 
     throw new \Exception("No data with the identifier {$id} was found.");
   }
 
+  /**
+   * {@inheritDoc}.
+   */
   public function retrieveAll(): array {
-    $node_nids = \Drupal::entityQuery('node')
+
+    $nodeStorage = $this->getNodeStorage();
+
+    $node_ids = $nodeStorage->getQuery()
       ->condition('type', $this->getType())
       ->condition('field_data_type', 'dataset')
       ->execute();
 
     $all = [];
-    foreach ($node_nids as $nid) {
-      $node = Node::load($nid);
+    foreach ($node_ids as $nid) {
+      $node = $nodeStorage->load($nid);
       $all[] = $this->themeDereferenced($node->field_json_metadata->value);
     }
     return $all;
   }
 
+  /**
+   * {@inheritDoc}.
+   */
   public function remove(string $id) {
 
-    foreach ($this->getNodesByUuid($id) as $nid) {
-      $node = Node::load($nid);
+    if (FALSE !== ($node = $this->getNodeByUuid($id))) {
       return $node->delete();
     }
   }
 
+  /**
+   * {@inheritDoc}.
+   */
   public function store(string $data, string $id = NULL): string {
 
     $data = json_decode($data);
@@ -65,42 +112,55 @@ class DrupalNodeDataset implements Storage, BulkRetriever {
     }
 
     if (!$id && isset($data->identifier)) {
-        $id = $data->identifier;
+      $id = $data->identifier;
     }
 
     if ($id) {
-        $node = \Drupal::service('entity.repository')->loadEntityByUuid('node', $id);
+      $node = $this->getNodeByUuid($id);
     }
 
     /* @var $node \Drupal\node\NodeInterface */
-    if ($node) {    // update existing node
+    // update existing node
+    if ($node) {
       $node->field_data_type = "dataset";
       $node->field_json_metadata = json_encode($data);
       $node->save();
       return $node->uuid();
     }
-    else {    // create new node
+    // Create new node.
+    else {
       $title = isset($data->title) ? $data->title : $data->name;
-      $nodeWrapper = NODE::create([
-        'title' => $title,
-        'type' => 'data',
-        'uuid' => $id,
-        'field_data_type' => 'dataset',
-        'field_json_metadata' => json_encode($data)
-      ]);
-      $nodeWrapper->save();
-      return $nodeWrapper->uuid();
+      $node = $this->getNodeStorage()
+        ->create([
+          'title' => $title,
+          'type' => 'data',
+          'uuid' => $id,
+          'field_data_type' => 'dataset',
+          'field_json_metadata' => json_encode($data),
+        ]);
+      $node->save();
+      return $node->uuid();
     }
 
     return NULL;
   }
 
-  private function getNodesByUuid($uuid) {
-    return \Drupal::entityQuery('node')
-      ->condition('type', $this->getType())
-      ->condition('field_data_type', 'dataset')
-      ->condition('uuid', $uuid)
-      ->execute();
+  /**
+   * Fetch node id of a current type given uuid.
+   *
+   * @return \Drupal\node\Entity\Node|bool
+   *   Returns false if no nodes match.
+   */
+  protected function getNodeByUuid($uuid) {
+
+    $nodes = $this->getNodeStorage()
+      ->loadByProperties([
+        'type' => $this->getType(),
+        'uuid' => $uuid,
+      ]);
+    // Uuid should be universally unique and always return
+    // a single node.
+    return current($nodes);
   }
 
   protected function themeDereferenced($json) {
